s390x: gs: move to new header file
Move the guarded-storage related structs and instructions to a new
header file because we will also need them for the SIGP store additional
status tests.
